dockerfiles/fastapi/requirements.txt
fastapi~=0.109
uvicorn~=0.27
boto3~=1.34
mlflow~=2.10
numpy~=1.24
pandas~=2.0
scikit-learn~=1.3
strawberry-graphql[fastapi]~=0.200.0
grpcio~=1.50.0
grpcio-tools~=1.50.0
kafka-python~=2.0.2